Airline and Airport now have a static init(std::istream&) method that initializes them, potentially asynchronously.
Flights has a factory method create() called by JS in fetch calls.
The JS passes Uint8Arrays of the data to the C++,
which then initializes Airlines and Airports
and then creates and returns the Flights object.

The Flights class doesn't have all the methods for accessing
the flights yet, but its construction and internal storage are done.

Moved some files:
    src/template to src/html since it's just pure html
    Map_image.svg to src/data