Got all the server side C++ code to work. Running ./CompactFlights reads in the CSV files, compacts them into Flights, and serializes them to flights.bin. It then checks to make sure it can properly deserialize flights.bin.
Running the analyzeFields() function in CompactFields.cpp (instead of compactFlights()),
does the analysis of the ranges and bits required for each field.
This matches the manual analysis I did earlier, so all the bit fields are correct.

I commented out the Wasm bindings for now to focus on the C++ server side,
but now I'll go back and compile it all to Wasm.

Apparently templates have to implemented in header files,
so I moved all template implementations to .tcc files,
which are included at the bottom of their corresponding header.
This resolves most of the undefined reference issues.

Set values for static fields in Airport and Airline
in .cpp files, which resolves the rest of the undefined reference issues.
To do so, I also had to create an empty constructor for DynamicEnum.

Fixed Date class, b/c mktime() doesn't convert day of year to time.
I had to do the efficient day of year to month and day of month myself,
using an algorithm adapted from http://howardhinnant.github.io/date_algorithms.html.

Adding std::ostream &operator<< to Date and Time for printing.

Added Time::between(Time, Time) to calculate the time between,
checking if the first time is actually after the second time,
meaning that the days are different, and then negating the result accordingly.
Do do this, I also overloaded all the comparison operators.