r3d3.js and wasm.js are correctly linked now. r3d3.js's main() functions are registered as postRun functions in Module, so they're run after the wasm code (it's empty main()) has finished running.
The JS Wasm interop isn't working yet,
but the bigger bug is in the Flask server.

I implemented a localStorage cachedFetch() method over fetch(),
although it doesn't work for large files like flights.bin,
since it exceeds localStorage's quota.

The reason I tried implementing this is that the Flask server
often doesn't receive the browser's request through fetch() for some reason.
I'm not sure if this an error in Chrome, Flask, or something on my computer.