Plotted all Airports.
Using d3.geoAlbersUsa(): GeoProjection,
I can map longitude, latitude to x, y on the SVG.
I adapted Queenie's code in index.js for drawing the states
using the d3.geoPath(projection), and then iterated through the airports,
creating circles at their locations with tooltips onmouseover
that display the airport name and IATA code.

I also reformulated the data loading flow.
The airlines, airports, and flights data is all loaded in parallel,
but the airlines and airports data is explicitly processed first
before the flights are loaded,
because the flights depend on the airlines and airports.
This was done in Flights.ts,
and I also had to rewrite the loading functions in C++.